example/requirements.txt
aws-mlops
boto3
GitPython
s3fs==2021.6.0
sagemaker
sklearn
pandas
